Create README.md
I designed a fan that uses a temperature and humidity sensor (DHT11 sensor) to control the motor speed to which the fan blades are attached. The 9V motor is powered by a 9V battery, and other components are powered by USB.
In addition, the LCD display shows the current temperature, the air humidity and allows manual motor speed control. The fan design is made with LEGO.

Components used:
 - Arduino UNO R3
 - DHT11 sensor
 - 16x2 LCD with I2C module
 - 9V DC  motor
 - 9V battery
 - 3x 10k ohm resistor
 - TIP41C transistor
 - 1N4007 diode
 - buzzer
 - G-55 DC 2.1
 - 2x tact switch buttons
 - 2x 400-Point breadboard 
 - USB cable